+++ packages/frontend/src/features/sqlRunner/components/ContentPanel.tsx
@@ -275,7 +275,6 @@ export const ContentPanel: FC = () => {
         config: currentVizConfig,
         sql,
         limit,
-        additionalQueryKey: [sql],
     });
 
     const [activeEchartsInstance, setActiveEchartsInstance] =